// src/circulargauge/src/circular-gauge/axes/axis-model.d.ts
export interface TickModel {

    /**
     * The width of the ticks in pixels.
     * @aspDefaultValueIgnore
     * @default null
     */

    width?: number;

    /**
     * The height of the line in pixels.
     * @aspDefaultValueIgnore
     * @default null
     */

    height?: number;

    /**
     * Specifies the interval of the tick line.
     * @aspDefaultValueIgnore
     * @default null
     */

    interval?: number;

    /**
     * Distance of the ticks from axis in pixel.
     * @default 0
     */

    offset?: number;

    /**
     * The color of the tick line, which accepts value in hex, rgba as a valid CSS color string.
     * @aspDefaultValueIgnore
     * @default null
     */

    color?: string;

    /**
     * Specifies the position of the ticks. They are
     * * inside -  Places the ticks inside the axis.
     * * outside - Places the ticks outside of the axis.
     * @default Inside
     */

    position?: Position;

    /**
     * If set true, major ticks takes the range color.
     * @default false
     */

    useRangeColor?: boolean;

}